backend/requirements.txt
fastapi
uvicorn[standard]
gspread
google-auth
python-dotenv
pydantic
fastapi-cors
